requirements.txt
arrow
steam
Pillow
lxml
translate
Shosetsu
imgurpython
praw
chatterbot
pyyaml
https://github.com/Rapptz/discord.py/archive/rewrite.zip#egg=discord.py[voice]
wikipedia
humanfriendly
python-twitch
wolframalpha
geopy
pymongo
google-api-python-client
pafy
youtube_dl
markovify
cssselect
soundcloud
ftfy
imageio
giphypop